Refactor: Formalise process pilot data function module structure
Configure the full url of the batch notify function as an env var. This loosens the coupling between the two functions.
