+++ src/functions/process_pilot_data/local.settings.json
@@ -4,6 +4,6 @@
     "FUNCTIONS_WORKER_RUNTIME": "python",
     "AzureWebJobsFeatureFlags": "EnableWorkerIndexing",
     "AzureWebJobsStorage": "UseDevelopmentStorage=true",
-    "BASE_URL": "http://localhost:7072"
+    "BATCH_NOTIFY_URL": "http://localhost:7072/api/batch-notify/breast-screening-pilot"
   }
 }
